requirements.txt
# Computer Vision and Image Processing
opencv-python==4.8.1.78
numpy>=1.21.0,<2.0.0
pillow>=9.0.0

# YOLO Object Detection
ultralytics==8.0.196

# PyTorch Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Additional useful packages for computer vision
matplotlib>=3.5.0
scipy>=1.8.0

# Configuration management
python-dotenv>=1.0.0
requests>=2.28.0

# CLIP for better object detection with text descriptions
git+https://github.com/openai/CLIP.git
Pillow>=9.0.0
PyYAML>=6.0

# Optional: For better performance and additional features
# psutil  # System monitoring
# tqdm    # Progress bars